2005-12-21  Peter Dennis Bartok  <pbartok@novell.com>
	* XplatUIWin32.cs: Fixed the TEXTMETRIC structure, we were calling
	  a Unicode function with a structure that was defined in Ansi way.
	  This fixes #76942.


svn path=/trunk/mcs/; revision=54720

+++ mcs/class/Managed.Windows.Forms/System.Windows.Forms/XplatUIWin32.cs
@@ -334,10 +334,10 @@ namespace System.Windows.Forms {
 			internal int			tmOverhang;
 			internal int			tmDigitizedAspectX;
 			internal int			tmDigitizedAspectY;
-			internal byte			tmFirstChar; 
-			internal byte			tmLastChar; 
-			internal byte			tmDefaultChar; 
-			internal byte			tmBreakChar; 
+			internal short			tmFirstChar; 
+			internal short			tmLastChar; 
+			internal short			tmDefaultChar; 
+			internal short			tmBreakChar; 
 			internal byte			tmItalic; 
 			internal byte			tmUnderlined; 
 			internal byte			tmStruckOut;
